cloud/admin-by-mistake/cloud-other-services/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
PyJWT==2.8.0
pycryptodome==3.10.1
requests==2.31.0
